requirements.txt
Flask>=3
requests>=2
python-dotenv>=1
gunicorn>=21
